Made DynamicIcon reuseable
