+++ components/specific/HRList.js
@@ -1,7 +1,5 @@
-import { germanDate } from '@/helpers/helpScripts';
+import { dynamicIconHandler, germanDate } from '@/helpers/helpScripts';
 import style from '@/layout/ContentLists.module.sass';
-import { faFile } from '@fortawesome/free-regular-svg-icons';
-import { faBuilding, faCircleMinus, faCirclePlus, faCodeBranch, faEllipsis, faGraduationCap, faUser } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 
@@ -12,23 +10,13 @@ const HRList = ({content}) => {
     return(
         <div>
             { content && content.data.map((item) => {
-                const iconMap = {
-                    'faEllipsis': faEllipsis,
-                    'faGraduationCap': faGraduationCap,
-                    'faUser': faUser,
-                    'faCirclePlus': faCirclePlus,
-                    'faBuilding': faBuilding,
-                    'faFile': faFile,
-                    'faCircleMinus': faCircleMinus,
-                    'faCodeBranch': faCodeBranch,
-                };
-                const icon = iconMap[item.attributes.pub_icon];
+                
                 return (
                     <Link href={'/hr/'+item.id} key={item.id}>
                         <div className={`${style.listItem} rounded-lg`}>
                             <div className={` ${style.listIcon} flex-none rounded-l-lg`}>
                                 <div className="w-5">
-                                {icon && <FontAwesomeIcon icon={icon} />}
+                                <FontAwesomeIcon icon={dynamicIconHandler(item.attributes.pub_icon)} />
                                 </div>
                             </div>
                             <div className={`${style.listContent} flex-auto`}>
